driver spec fortran kind

+++ sample/FORTRAN/driver_spec.f90
@@ -32,10 +32,10 @@
       real(mytype), dimension(:,:,:),  allocatable :: BEG
       complex(mytype), dimension(:,:,:),  allocatable :: AEND
 
-      integer(8) Ntot
+      integer(i8) Ntot
       real(mytype) factor
-      real(8) rtime1,rtime2
-      real(8) gt(4,3),gtcomm(3),tc
+      real(r8) rtime1,rtime2
+      real(r8) gt(4,3),gtcomm(3),tc
       integer ierr,nu,ndim,dims(2),nproc,proc_id
       integer istart(3),iend(3),isize(3)
       integer fstart(3),fend(3),fsize(3)
@@ -319,7 +319,7 @@
 
       use p3dfft
 
-      integer(8) nar,i
+      integer(i8) nar,i
       complex(mytype) X(nar)
       real(mytype) f
